Preliminary Socket implementation and event rework.
Socket implementation still does not work, many needed classes are there and some TODOs will be solved in the following days. VirtualView and NetworkHandler are not yet fully implemented.

Events have been slightly reworked but not yet implemented to comply with lab5 architectural choices shown in the networking slide. Mainly Events have been divided up in two VCEvent (View to Controller) and MVEvent (Model to View event)

UML is not up-to-date with those changes, it will be adjusted when a working and tested implementation of the network part will be in place.